+++ Code/Source/Clients/SpriteRequestHandler.h
@@ -38,6 +38,7 @@ namespace Diorama
 
         // DioramaSpriteRequests
         bool SetTextureByPath(AZStd::string_view productPath) override;
+        bool SetNormalMapByPath(AZStd::string_view productPath) override;
         void SetSize(float width, float height) override;
         void SetPivot(float x, float y) override;
         void SetTint(float r, float g, float b, float a) override;